refactor and rename executor -> execute
