+++ workers/rust/container/src/task.rs
@@ -1,5 +1,5 @@
 use serde::Deserialize;
-use taskcluster_lib_worker::executor;
+use taskcluster_lib_worker::execute;
 
 /// A container-worker payload
 #[derive(Debug, Deserialize)]
@@ -8,7 +8,7 @@ pub(crate) struct Payload {
     pub(crate) command: Vec<String>,
 }
 
-impl executor::Payload for Payload {
+impl execute::Payload for Payload {
     fn from_value(v: serde_json::Value) -> Result<Self, anyhow::Error> {
         Ok(serde_json::from_value(v)?)
     }
