+++ workers/rust/lib-worker/src/claim/work_claimer.rs
@@ -1,5 +1,5 @@
 use super::long_poll::{ClaimWorkLongPoll, LongPollCommand};
-use crate::executor::{ExecutionFactory, Executor, Payload};
+use crate::execute::{ExecutionFactory, Executor, Payload};
 use crate::process::{Process, ProcessFactory, ProcessSet};
 use anyhow::{bail, Context as AnyhowContext, Result};
 use async_trait::async_trait;
