+++ workers/rust/lib-worker/src/execute/creds_container.rs
@@ -1,0 +1,51 @@
+use crate::execute::QueueFactory;
+use std::sync::{Arc, Mutex};
+use taskcluster::{ClientBuilder, Credentials, Queue};
+
+/// Clonable credentials container, allowing updates as they expire.  This is used as the
+/// QueueFactory.
+#[derive(Clone)]
+pub(super) struct CredsContainer(Arc<Mutex<Inner>>);
+
+struct Inner {
+    root_url: String,
+    creds: Credentials,
+    queue: Option<Arc<Queue>>,
+}
+
+#[allow(dead_code)]
+impl CredsContainer {
+    pub(super) fn new(root_url: String, creds: Credentials) -> Self {
+        Self(Arc::new(Mutex::new(Inner {
+            root_url,
+            creds,
+            queue: None,
+        })))
+    }
+
+    pub(super) fn get(&self) -> Credentials {
+        return self.0.lock().unwrap().creds.clone();
+    }
+
+    pub(super) fn set(&self, creds: Credentials) {
+        let mut inner = self.0.lock().unwrap();
+        inner.creds = creds;
+        // queue is invalidated, so reset it to None
+        inner.queue = None;
+    }
+}
+
+impl QueueFactory for CredsContainer {
+    fn queue(&self) -> anyhow::Result<Arc<Queue>> {
+        let mut inner = self.0.lock().unwrap();
+        if let Some(ref queue) = inner.queue {
+            Ok((*queue).clone())
+        } else {
+            let queue = Arc::new(Queue::new(
+                ClientBuilder::new(&inner.root_url).credentials(inner.creds.clone()),
+            )?);
+            inner.queue = Some(queue.clone());
+            Ok(queue)
+        }
+    }
+}
