+++ workers/rust/lib-worker/src/lib.rs
@@ -2,15 +2,15 @@
 //! supplementing the Rust client.
 //!
 //! * [`claim`] implements work claiming
-//! * [`executor`] abstracts task execution
+//! * [`execute`] abstracts task execution
 //! * [`process`] provides an abstraction for concurrent proceses
 //! * [`task`] provides a deserializable task struct, with some useful utilities
 //!
 //! A typical worker will start a [`claim::WorkClaimer`], passing it a custom
-//! [`executor::Executor`] implementation.
+//! [`execute::Executor`] implementation.
 
 pub mod claim;
-pub mod executor;
+pub mod execute;
 pub mod process;
 pub mod task;
